reflect changes to lwmqtt

+++ esp_mqtt.h
@@ -17,7 +17,7 @@ typedef void (*esp_mqtt_status_callback_t)(esp_mqtt_status_t);
 /**
  * The message callback.
  */
-typedef void (*esp_mqtt_message_callback_t)(const char *topic, const char *payload, unsigned int len);
+typedef void (*esp_mqtt_message_callback_t)(const char *topic, const char *payload, int len);
 
 /**
  * Initialize the MQTT management system.
@@ -30,7 +30,7 @@ typedef void (*esp_mqtt_message_callback_t)(const char *topic, const char *paylo
  * @param command_timeout - The command timeout.
  */
 void esp_mqtt_init(esp_mqtt_status_callback_t scb, esp_mqtt_message_callback_t mcb, int buffer_size,
-                   unsigned int command_timeout);
+                   int command_timeout);
 
 /**
  * Start the MQTT process.
@@ -41,7 +41,7 @@ void esp_mqtt_init(esp_mqtt_status_callback_t scb, esp_mqtt_message_callback_t m
  * @param username - The client username.
  * @param password - The client password.
  */
-void esp_mqtt_start(const char *host, unsigned int port, const char *client_id, const char *username,
+void esp_mqtt_start(const char *host, int port, const char *client_id, const char *username,
                     const char *password);
 
 /**
@@ -71,7 +71,7 @@ bool esp_mqtt_unsubscribe(const char *topic);
  * @param retained - The retained flag.
  * @return
  */
-bool esp_mqtt_publish(const char *topic, void *payload, uint16_t len, int qos, bool retained);
+bool esp_mqtt_publish(const char *topic, void *payload, int len, int qos, bool retained);
 
 /**
  * Stop the MQTT process.
